Extract CE logic into hooks
